requirements.txt
numpy 
matplotlib
sendgrid
python-dotenv
smac
